make RoomDetailRow reusable for the Room Directory
